requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
django-celery-beat==2.5.0
celery==5.3.4
psycopg2-binary==2.9.9
python-decouple==3.8
PyJWT==2.8.1
django-rest-framework-simplejwt==5.3.2
Pillow==10.1.0
django-environ==0.21.0
djangorestframework-pdf==0.1.2
reportlab==4.0.7
django-extensions==3.2.3
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
